+++ src/DockCraft/general/__init__.py
@@ -1,1 +1,7 @@
+"""
+the general module contains the general classe
+and functions used by the DockCraft package
+"""
+
 from .dockcraft_reader import DockerfileReader
+from .dockcraft_file_finder import DockCraftFinder