+++ CHANGELOG.md
@@ -4,9 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2024-08-14
+### Added
+- Feature: DockCraft File Finder
+### Changed
+- Changed the main pipeline:
+    - added the DockCraft File Finder to the init
+    - added the DockCraft Reader to the init
+
 ## [0.2.0] - 2024-08-14
 ### Added
-- Feature: DockerCraft Reader
+- Feature: DockCraft Reader
 - Created the first iteration of the main pipeline
 ### Removed
 - Removed mock tests for Pipeline Setup